+++ .travis.yml
@@ -12,10 +12,9 @@ install:
 before_script:
 - rm $HOME/.m2/settings.xml
 - export MAVEN_OPTS="-Dmaven.artifact.threads=3 -Xmx2469m -XX:MaxPermSize=512m"
-- export PROJECTS="--projects !bundles/automation/org.eclipse.smarthome.automation.module.timer.test/pom.xml,!extensions/voice/org.eclipse.smarthome.voice.mactts/pom.xml,!bundles/automation/org.eclipse.smarthome.automation.module.script.test/pom.xml"
 
 script:
-- MAVEN_SKIP_RC=true apache-maven-3.3.9/bin/mvn install -fae -B $PROJECTS 2>&1 | ./filter.py ; test ${PIPESTATUS[0]} -eq 0
+- MAVEN_SKIP_RC=true apache-maven-3.3.9/bin/mvn install -fae -B 2>&1 | ./filter.py ; test ${PIPESTATUS[0]} -eq 0
 
 addons:
   apt:
