+++ bundles/io/org.eclipse.smarthome.io.voice/.classpath
@@ -10,7 +10,5 @@
 	<classpathentry kind="con" path="org.eclipse.pde.core.requiredPlugins"/>
 	<classpathentry kind="src" path="src/main/java"/>
 	<classpathentry kind="src" path="src/main/resources"/>
-	<classpathentry kind="lib" path="lib/native-lib-loader-2.1.3.jar"/>
-	<classpathentry kind="lib" path="lib/nar-maven-pocketsphinx-1.0-SNAPSHOT.jar"/>
 	<classpathentry kind="output" path="target/classes"/>
 </classpath>
