+++ bundles/io/org.eclipse.smarthome.io.voice/META-INF/MANIFEST.MF
@@ -44,9 +44,7 @@ Import-Package: javax.sound.sampled,
  org.eclipse.smarthome.io.voice,
  org.osgi.framework,
  org.slf4j
-Export-Package: org.eclipse.smarthome.io.voice.text,
- org.eclipse.smarthome.io.voice.tts,
- org.eclipse.smarthome.io.voice.stt,
- org.eclipse.smarthome.io.voice
+Export-Package: org.eclipse.smarthome.io.voice,
+ org.eclipse.smarthome.io.voice.text
 Service-Component: OSGI-INF/*.xml
 Bundle-Activator: org.eclipse.smarthome.io.voice.internal.VoiceActivator
