--- bundles/io/org.eclipse.smarthome.io.voice/OSGI-INF/TTSServiceMacOS.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Copyright (c) 2014-2015 openHAB UG (haftungsbeschraenkt) and others.
-    All rights reserved. This program and the accompanying materials
-    are made available under the terms of the Eclipse Public License v1.0
-    which accompanies this distribution, and is available at
-    http://www.eclipse.org/legal/epl-v10.html
-
--->
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" immediate="true" name="org.eclipse.smarthome.io.voice.mactts">
-   <implementation class="org.eclipse.smarthome.io.voice.internal.TTSServiceMacOS"/>
-   <service>
-      <provide interface="org.eclipse.smarthome.io.voice.TTSService"/>
-   </service>
-   <property name="os" type="String" value="macosx"/>
-</scr:component>
