+++ extensions/voice/org.eclipse.smarthome.voice.mactts.test/.classpath
@@ -8,6 +8,6 @@
 	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.7"/>
 	<classpathentry kind="con" path="org.eclipse.pde.core.requiredPlugins"/>
 	<classpathentry kind="src" path="src/test/java"/>
-	<classpathentry combineaccessrules="false" kind="src" path="/org.eclipse.smarthome.io.voice"/>
+	<classpathentry combineaccessrules="false" kind="src" path="/org.eclipse.smarthome.voice.mactts"/>
 	<classpathentry kind="output" path="target/classes"/>
 </classpath>
