+++ extensions/voice/org.eclipse.smarthome.voice.mactts.test/.settings/org.eclipse.pde.core.prefs
@@ -1,0 +1,4 @@
+#Mon Oct 11 21:08:09 CEST 2010
+eclipse.preferences.version=1
+pluginProject.extensions=false
+resolve.requirebundle=false
