+++ extensions/voice/org.eclipse.smarthome.voice.mactts.test/META-INF/MANIFEST.MF
@@ -1,0 +1,11 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: Eclipse SmartHome Mac TTS Tests
+Bundle-SymbolicName: org.eclipse.smarthome.voice.mactts.test
+Bundle-Version: 0.9.0.qualifier
+Bundle-Vendor: Eclipse.org/SmartHome
+Fragment-Host: org.eclipse.smarthome.voice.mactts
+Bundle-RequiredExecutionEnvironment: JavaSE-1.7
+Import-Package: org.junit;version="4.0.0",
+ org.eclipse.smarthome.voice.mactts
+Bundle-ClassPath: .
