+++ extensions/voice/org.eclipse.smarthome.voice.mactts.test/build.properties
@@ -1,0 +1,5 @@
+source.. = src/test/java/
+output.. = target/test-classes/
+bin.includes = META-INF/,\
+               .,\
+               about.html
