+++ extensions/voice/org.eclipse.smarthome.voice.mactts.test/src/test/java/org/eclipse/smarthome/voice/mactts/internal/VoiceMacOSTest.java
@@ -5,7 +5,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.eclipse.smarthome.io.voice.internal;
+package org.eclipse.smarthome.voice.mactts.internal;
 
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
