+++ extensions/voice/org.eclipse.smarthome.voice.mactts/META-INF/MANIFEST.MF
@@ -6,6 +6,8 @@ Bundle-Vendor: Eclipse.org/SmartHome
 Bundle-Version: 0.9.0.qualifier
 Bundle-RequiredExecutionEnvironment: JavaSE-1.7
 Bundle-ClassPath: .
-Import-Package: org.eclipse.smarthome.io.voice.tts,
+Import-Package: org.eclipse.smarthome.io.voice,
+ org.eclipse.smarthome.io.audio,
  org.slf4j
+Export-Package: org.eclipse.smarthome.voice.mactts
 Service-Component: OSGI-INF/*.xml
