+++ extensions/voice/org.eclipse.smarthome.voice.mactts/OSGI-INF/TTSServiceMacOS.xml
@@ -8,9 +8,10 @@
     http://www.eclipse.org/legal/epl-v10.html
 
 -->
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" immediate="true" name="org.eclipse.smarthome.io.voice.mac">
-   <implementation class="org.eclipse.smarthome.io.voice.internal.TTSServiceMacOS"/>
+<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" immediate="true" name="org.eclipse.smarthome.voice.mactts.mac">
+   <implementation class="org.eclipse.smarthome.voice.mactts.internal.TTSServiceMacOS"/>
    <service>
       <provide interface="org.eclipse.smarthome.io.voice.TTSService"/>
    </service>
+   <property name="os" type="String" value="macosx"/>
 </scr:component>
